+++ pypaimon/api/table_commit.py
@@ -1,0 +1,36 @@
+################################################################################
+#  Licensed to the Apache Software Foundation (ASF) under one
+#  or more contributor license agreements.  See the NOTICE file
+#  distributed with this work for additional information
+#  regarding copyright ownership.  The ASF licenses this file
+#  to you under the Apache License, Version 2.0 (the
+#  "License"); you may not use this file except in compliance
+#  with the License.  You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+# limitations under the License.
+#################################################################################
+
+from abc import ABC, abstractmethod
+from pypaimon.api import CommitMessage
+from typing import List
+
+
+class BatchTableCommit(ABC):
+    """A table commit for batch processing. Recommended for one-time committing."""
+
+    @abstractmethod
+    def commit(self, commit_messages: List[CommitMessage]):
+        """
+        Commit the commit messages to generate snapshots. One commit may generate
+        up to two snapshots, one for adding new files and the other for compaction.
+        """
+
+    @abstractmethod
+    def close(self):
+        """Close this resource."""
