+++ tox.ini
@@ -45,6 +45,6 @@ max-line-length=100
 exclude=.tox/*,dev/*,build/*,dist/*
 
 [mypy]
-files=paimon_python_api/*.py
+files=pypaimon/api/*.py
 ignore_missing_imports = True
 strict_optional=False
